test: non-null assertions output
